requirements.txt
PyYAML==6.0.1
boto3==1.34.4
cryptography==41.0.7
requests==2.31.0
